backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
psycopg2-binary
pydantic[email]
python-dotenv
python-multipart
cryptography
passlib[bcrypt]
python-jose[cryptography]
